Moving utils => __init__ in submodules
