+++ src/qumin/representations/__init__.py
@@ -1,3 +1,167 @@
 # -*- coding: utf-8 -*-
-# !/usr/bin/env python3
-from .utils import *
+# !/usr/bin/env python
+"""author: Sacha Beniamine.
+
+Utility functions for representations.
+"""
+import logging
+from collections import defaultdict
+
+import numpy as np
+import pandas as pd
+
+from .segments import Inventory, Form
+from ..utils import merge_duplicate_columns
+
+log = logging.getLogger()
+
+
+def unique_lexemes(series, file_type):
+    """Check if there are duplicates in the lexemes list.
+    If yes, raise an error.
+
+    Arguments:
+        series (:class:`pandas:pandas.Series`): list of lexemes as a Series object.
+        file_type (str): used for error messages. Should describe the dataset which is being tested.
+    """
+    duplicated = list(series[series.duplicated()])
+
+    if duplicated:
+        raise ValueError(f"""There are {len(duplicated)} duplicates among lexemes.
+            Please check the {file_type} table.
+            Duplicates are: {", ".join(duplicated)}""")
+
+
+def create_features(md, feature_cols):
+    """Read feature and preprocess to be coindexed with paradigms."""
+    if type(feature_cols) == str:
+        feature_cols = [feature_cols]
+    features = pd.read_csv(md.get_table_path('lexemes'))
+    unique_lexemes(features["lexeme_id"], "lexemes")
+    features.set_index("lexeme_id", inplace=True)
+    features.fillna(value="", inplace=True)
+    return features.loc[:, feature_cols]
+
+
+def create_paradigms(data_file_name, verbose=False, fillna=True,
+                     segcheck=False, merge_duplicates=False,
+                     defective=False, overabundant=False, merge_cols=False,
+                     cells=None):
+    """Read paradigms data, and prepare it according to a Segment class pool.
+
+    Arguments:
+        data_file_name (str): path to the paradigm csv file.
+        All characters occuring in the paradigms except the first column
+        should be inventoried in this class.
+        verbose (bool): verbosity switch.
+        merge_duplicates (bool): should identical columns be merged ?
+        fillna (bool): Defaults to True. Should #DEF# be replaced by np.NaN ? Otherwise they are filled with empty strings ("").
+        segcheck (bool): Defaults to False. Should I check that all the phonological segments in the table are defined in the segments table ?
+        defective (bool): Defaults to False. Should I keep rows with defective forms ?
+        overabundant (bool): Defaults to False. Should I keep rows with overabundant forms ?
+        merge_cols (bool): Defaults to False. Should I merge identical columns (fully syncretic) ?
+        cells (List[str]): List of cell names to consider. Defaults to all.
+
+    Returns:
+        paradigms (:class:`pandas:pandas.DataFrame`): paradigms table (columns are cells, index are lemmas).
+    """
+
+    def get_unknown_segments(forms, unknowns, name):
+        known_sounds = set(Inventory._classes) | set(Inventory._normalization) | {";", ""}
+        for form in forms.split(";"):
+            if " " in form:
+                tokens = form.split(" ")
+            else:
+                tokens = Inventory._segmenter.split(form)
+            for char in tokens:
+                if char not in known_sounds:
+                    unknowns[char].append(form + " " + name)
+
+    # Reading the paradigms.
+    lexemes, cell_col, form_col = ("lexeme", "cell", "phon_form")
+    paradigms = pd.read_csv(data_file_name, na_values=["", "#DEF#"], dtype="str", keep_default_na=False,
+                            usecols=["form_id", lexemes, cell_col, form_col])
+    paradigms = paradigms.drop_duplicates([lexemes, cell_col, form_col])
+
+    def aggregator(s):
+        if s.isnull().all():
+            return None
+        return ";".join(s.values)
+
+    def check_cells(cells, par_cols):
+        unknown_cells = set(cells) - set(par_cols)
+        if unknown_cells:
+            raise ValueError(f"You specified some cells which aren't in the paradigm : {' '.join(unknown_cells)}")
+        return sorted(list(set(par_cols) - set(cells)))
+
+    if not {lexemes, cell_col, form_col} < set(paradigms.columns):
+        log.warning("Please use Paralex-style long-form table (http://www.paralex-standard.org).")
+
+    # Filter cells before pivoting for speed reasons
+    if cells is not None:
+        to_drop = check_cells(cells, paradigms[cell_col].unique())
+        if len(to_drop) > 0:
+            log.info(f"Dropping rows with following cell values: {', '.join(sorted(to_drop))}")
+            paradigms = paradigms[(paradigms[cell_col].isin(cells))]
+
+    paradigms = paradigms.pivot_table(values=form_col, index=lexemes,
+                                      columns=cell_col,
+                                      aggfunc=aggregator)
+
+    paradigms.reset_index(inplace=True, drop=False)
+
+    if not defective:
+        paradigms.dropna(axis=0, inplace=True)
+
+    log.debug(paradigms)
+
+    # Lexemes must be unique identifiers
+    unique_lexemes(paradigms[lexemes], 'paradigms')
+    paradigms.set_index(lexemes, inplace=True)
+
+    paradigms.fillna(value="", inplace=True)
+    if merge_duplicates:
+        agenda = list(paradigms.columns)
+        while agenda:
+            a = agenda.pop(0)
+            for i, b in enumerate(agenda):
+                if (paradigms[a] == paradigms[b]).all():
+                    log.debug("Identical columns %s and %s ", a, b)
+                    new = a + " & " + b
+                    agenda.pop(i)
+                    agenda.append(new)
+                    paradigms[new] = paradigms[a]
+                    paradigms.drop([a, b], inplace=True, axis=1)
+                    break
+
+    if segcheck:
+        log.info("Checking we have definitions for all the phonological segments in this data...")
+        unknowns = defaultdict(list)
+        paradigms.apply(lambda x: x.apply(get_unknown_segments, args=(unknowns, x.name)), axis=1)
+
+        if len(unknowns) > 0:
+            alert = "Your paradigm has unknown segments: " + "\n ".join(
+                "[{}] (in {} forms:{}) ".format(u, len(unknowns[u]), ", ".join(unknowns[u][:10])) for u in unknowns)
+            raise ValueError(alert)
+
+    def parse_cell(cell):
+        if not cell:
+            return cell
+        forms = [Form(f) for f in cell.split(";")]
+        if overabundant:
+            forms = tuple(sorted(forms))
+        else:
+            forms = (forms[0],)
+        return forms
+
+    paradigms = paradigms.map(parse_cell)
+
+    log.info("Merging identical columns...")
+    if merge_cols:
+        merge_duplicate_columns(paradigms, sep="#")
+
+    if not fillna:
+        paradigms = paradigms.replace("", np.NaN)
+
+    log.debug(paradigms)
+    return paradigms
